+++ CHANGELOG.md
@@ -26,7 +26,8 @@ This version add a main dashboard with main financial statistics
 - Added logic for no pending waiter orders in `src/templates/orders/waiter/cards.html`
 
 ### Fixed
-- N/A
+- Entire flow to calculate totals add discounts and tip correctly in `src/api/v1/apps/orders/services.py`
+- Fix the correct visualization of totals and discounts in `src/templates/orders/check_out.html`
 
 ## [2.0.0][Released] 2026-06-15
 First release with a MVP of herenciapp, with all the customer-kitchen-waiter-checkout flow